refactor(BAlert): namespace out useCountdown
