+++ packages/bootstrap-vue-next/src/components/BAlert/BAlert.vue
@@ -31,8 +31,7 @@ import BCloseButton from '../BButton/BCloseButton.vue'
 import BButton from '../BButton/BButton.vue'
 import type {Booleanish, ColorVariant} from '../../types'
 import {computed, onBeforeUnmount, toRef, useSlots, watchEffect} from 'vue'
-import {useBooleanish} from '../../composables'
-import useCountdown from '../../composables/useCountdown'
+import {useBooleanish, useCountdown} from '../../composables'
 import {isEmptySlot} from '../../utils'
 import {useVModel} from '@vueuse/core'
