+++ packages/bootstrap-vue-next/src/composables/index.ts
@@ -1,6 +1,8 @@
 export {default as useAlignment} from './useAlignment'
 export {default as useBooleanish} from './useBooleanish'
 export {createBreadcrumb, useBreadcrumb} from './useBreadcrumb'
+export {default as useColorMode} from './useColorMode'
+export {default as useCountdown} from './useCountdown'
 export {
   bindGroupProps,
   getClasses,
@@ -11,7 +13,6 @@ export {
   optionToElement,
   slotsToElements,
 } from './useFormCheck'
-export {default as useFormInput, COMMON_INPUT_PROPS} from './useFormInput'
+export {COMMON_INPUT_PROPS, default as useFormInput} from './useFormInput'
 export {normalizeOptions} from './useFormSelect'
 export {default as useId} from './useId'
-export {default as useColorMode} from './useColorMode'
